airflow_docker/requirements.txt
async-timeout==4.0.3
configparser==6.0.0
importlib-resources==6.1.1
numpy==1.24.4
openpyxl==3.0.3
pandas==2.0.3
redshift-connector==2.1.0
SQLAlchemy==1.4.51
sqlalchemy-redshift==0.8.14
openmeteo-requests==1.2.0
requests-cache==1.2.0
retry-requests==2.0.0
psycopg2-binary==2.9.9
requests==2.31.0